Moved to AddNewCard component
